fix image format for libfprint

+++ libfprint/drivers/goodix/5395/goodix_protocol.h
@@ -50,7 +50,7 @@ void fpi_goodix_protocol_encode(GoodixMessage *message, gboolean calc_checksum, 
 gboolean fpi_goodix_protocol_decode(guint8 *data, GoodixMessage **message, GError **error);
 gboolean fpi_goodix_protocol_check_ack(GoodixMessage *message, GError **error);
 int fpi_goodix_protocol_decode_u32(guint8 *data, uint length);
-GoodixMessage *fpi_goodix_protocol_create_message(guint8 category, guint8 command, guint8 *payload, guint8 length);
+GoodixMessage *fpi_goodix_protocol_create_message(guint8 category, guint8 command, guint8 *payload, guint32 length);
 GoodixMessage *fpi_goodix_protocol_create_message_byte_array(guint8 category, guint8 command, GByteArray *payload);
 GArray* fpi_goodix_protocol_decode_image(const GByteArray *image);
 gboolean fpi_goodix_protocol_verify_otp_hash(const guint8 *otp, guint otp_length, const guint8 otp_hash[]);
